moved wordsService test to own file, improved it
